パフォーマンス定まってきた

+++ main_board/src/led/led.h
@@ -9,7 +9,7 @@ class LED{
         LED_PARTS led_parts = LED_PARTS();
         UTILITY utility = UTILITY();
 
-        double i = 0;
+        float i = 0;
         int panel_A_ID = 1000; int panel_B_ID = 1000; int panel_C_ID = 1000; int panel_N = 1000;
         int panel_ver = 48; int panel_hor = 48; //RGBのマックス,パネルの縦横LED数
         int px, py;
